--- app/routes/_app.alchemy.converting._index/_index.tsx
@@ -1,10 +1,0 @@
-export default function AlchemyConvertingEmptyPage() {
-  return (
-    <div className="text-center">
-      <h2 className="mt-2 text-sm font-medium text-gray-12">Convert tab</h2>
-      <p className="mt-1 text-sm text-gray-11">
-        Click item on the item you want to convert
-      </p>
-    </div>
-  )
-}
