+++ app/routes/_app.alchemy.converting/_index.tsx
@@ -1,13 +1,10 @@
 import * as RemixNode from '@remix-run/node'
 import * as RemixReact from '@remix-run/react'
-import clsx from 'clsx'
-import Image from '~/components/Image'
-import Tooltip from '~/components/Tooltip'
 import * as ItemData from '~/data/items'
 import * as DB from '~/db.server'
 import { getInventoryByType } from '~/models/inventory.server'
 import * as Session from '~/session.server'
-import * as Utils from '~/utils'
+import Item from './Item'
 
 export async function loader({ request }: RemixNode.LoaderArgs) {
   const accountId = await Session.requireAccountId(request)
@@ -21,12 +18,13 @@ export async function loader({ request }: RemixNode.LoaderArgs) {
     ...ItemData.getItemsByType({
       type: DB.ItemType.ASCENSION_GEM,
       userItems: inventory,
-    }),
+    }).filter((item) => !item.name.includes('Brilliant Diamond')),
+
     ...ItemData.getItemsByType({
       type: DB.ItemType.TALENT_BOSS,
       userItems: inventory,
     }),
-  ].filter((item) => !item.name.includes('Brilliant Diamond'))
+  ]
 
   return RemixNode.json({ items })
 }
@@ -34,68 +32,48 @@ export async function loader({ request }: RemixNode.LoaderArgs) {
 export default function AlchemyConvertingPage() {
   const { items } = RemixReact.useLoaderData<typeof loader>()
 
-  return (
-    <div className="mt-8 grid gap-8 sm:grid-cols-2">
-      <div className="col-span-1 flex flex-wrap items-center gap-4">
-        {items.map((item) => (
-          <Item
-            key={item.name}
-            name={item.name}
-            rarity={item.rarity}
-            quantity={item.quantity ?? 0}
-          />
-        ))}
-      </div>
-      <div className="col-span-1">
-        <RemixReact.Outlet />
-      </div>
-    </div>
+  const ascensionGemItems = items.filter(
+    (item) => item.type === 'ASCENSION_GEM'
   )
-}
+  const talentBossItems = items.filter((item) => item.type === 'TALENT_BOSS')
 
-const backgroundImage: Record<1 | 2 | 3 | 4 | 5, string> = {
-  1: 'bg-image-rarity-1',
-  2: 'bg-image-rarity-2',
-  3: 'bg-image-rarity-3',
-  4: 'bg-image-rarity-4',
-  5: 'bg-image-rarity-5',
-}
-
-function Item({
-  name,
-  rarity,
-  quantity,
-}: {
-  name: string
-  rarity: 1 | 2 | 3 | 4 | 5
-  quantity: number
-}) {
   return (
-    <RemixReact.Link prefetch="intent" to={`./${name}`}>
-      <Tooltip text={name}>
-        <div className={clsx('rounded-b-md bg-gray-3 shadow-sm')}>
-          <div
-            className={clsx(
-              backgroundImage[rarity],
-              'rounded-t-md rounded-br-2xl bg-contain'
-            )}
-          >
-            <Image
-              src={`/item/${Utils.getImageSrc(name)}.png`}
-              alt={name}
-              className="h-16 w-16 rounded-br-2xl"
-              width={64}
-              height={64}
-            />
-          </div>
-          <div className="text-center">
-            <span className="sr-only">Quantity {quantity}</span>
-            <p className="text-sm text-gray-11" aria-hidden>
-              {quantity}
-            </p>
+    <>
+      <div className="mt-8 grid gap-8 sm:grid-cols-2">
+        <div className="col-span-1">
+          <h2 className="text-lg font-medium leading-6 text-gray-12">
+            Ascension Gems
+          </h2>
+          <div className="mt-6 flex flex-wrap gap-4">
+            {ascensionGemItems.map((item) => (
+              <Item
+                key={item.name}
+                type="gem"
+                name={item.name}
+                rarity={item.rarity}
+                quantity={item.quantity ?? 0}
+              />
+            ))}
           </div>
         </div>
-      </Tooltip>
-    </RemixReact.Link>
+        <div className="col-span-1 self-start">
+          <h2 className="text-lg font-medium leading-6 text-gray-12">
+            Talent Boss
+          </h2>
+          <div className="mt-6 flex flex-wrap gap-4">
+            {talentBossItems.map((item) => (
+              <Item
+                key={item.name}
+                type="boss"
+                name={item.name}
+                rarity={item.rarity}
+                quantity={item.quantity ?? 0}
+              />
+            ))}
+          </div>
+        </div>
+      </div>
+      <RemixReact.Outlet />
+    </>
   )
 }
